Update JUnit (4.8->4.9), refactor, remove HibernateSpringTxMethodRule.resetThreadsForCurrentTestClass(), hide HibernateSpringTxMethodRule.createSessionAndBeginTransaction(), document, add travis ci
Update JUnit to gain access to TestRule - a better replacement for MethodRule.
Refactor by replacing maps in HibernateSpringTxMethodRule with ThreadLocals.
Remove HibernateSpringTxMethodRule.resetThreadsForCurrentTestClass() method. It was a very bad design to require that the user call it to clean up after a test. Now it is automaticly detected that a test class has changed for the current thread.
There was no need to expose HibernateSpringTxMethodRule.createSessionAndBeginTransaction() so it was hidden.
Improve documentation.