Add "lazy" sessionFactory and hibernateTemplate
"Lazy" sessionFactory is returned by HibernateSpringTxMethodRule.getSessionFactory(). It must be lazy because this method can be invoked before the rule is applied (and the actual sessionFactory is created when the rule is applied), for example if a dao is a field of a test class and it is initialized (dao=new DaoImpl(rule.getSessionFactory()) in a constructor of the test class and not in a "before" method (in "before" method the rule is already applied). Lazy hibernateTemplate exists for the same reason.